requirements.txt
Telethon
PyQt6
qasync
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
